refactor: group main routes + rename search
