+++ src/components/landing/landing-header/search-input.tsx
@@ -1,6 +1,6 @@
 import { Input } from "@/components/ui/input";
 import { SearchIcon } from "lucide-react";
-import BookSearch from "@/components/book-search/search";
+import BookSearch from "@/components/book-search/book-search";
 
 export default function SearchInput() {
   // TODO: Add search functionality
